Add fully working initial implementation
Still need to iron out a few details, for example a cleanup in what data is sent where is needed. Maybe user_id should not be required in the wsconnection, but only in the client messages. Conversely, we should drop the race_id from the client messages, as the client is already connected to a race instance