Websocket server (#6)
* Add initial progress

* Add copiling and running but very scuffed websocket implementation

* Add semi-working ws implementation

I'm now able to propagate messages from one client to another, but the second client's messages are not passed to the initial connection. Connecting several clients results in similar behavious, where only the first client sends messages to the last, and any clients in between are left dormant

* Add fully working initial implementation

Still need to iron out a few details, for example a cleanup in what data is sent where is needed. Maybe user_id should not be required in the wsconnection, but only in the client messages. Conversely, we should drop the race_id from the client messages, as the client is already connected to a race instance

* Implement heartbeat checks

* Remove unnecessary Dockerfile contents

* Handle invalid ClientActorMessage parsing

* Remove unused imports & improve some logging

* Refactor ClientActorMessage into LocationUpdateMessage and improve error handling

* Update app with user and race input fields

* Update port and set liveness checks logging to debug

* Add caching layer to Dockerfile

* disable input field capitalization and autocorrect

* Fix Dockerfile

* Refactor ws message handling + implement binary message support

* Remove unused imports

* Fix invalid payload and change request url

* Rename Coordinates struct fields and add some logging